fix: Fix Search Dialog Display switch Sidebar Status - Meeds-io/MIPs#175
