+++ webapp/src/main/webapp/vue-apps/search/initComponents.js
@@ -1,5 +1,6 @@
 import SearchApplication from './components/SearchApplication.vue';
 import SearchToolbar from './components/SearchToolbar.vue';
+import SearchDialog from './components/SearchDialog.vue';
 import SearchResults from './components/SearchResults.vue';
 import SearchResultCard from './components/SearchResultCard.vue';
 import SearchTagSelector from './components/SearchTagSelector.vue';
@@ -8,6 +9,7 @@ import SearchTagList from './components/SearchTagList.vue';
 const components = {
   'search-application': SearchApplication,
   'search-toolbar': SearchToolbar,
+  'search-dialog': SearchDialog,
   'search-results': SearchResults,
   'search-result-card': SearchResultCard,
   'search-tag-selector': SearchTagSelector,
